requirements.txt
Django==1.11.17
patool==1.12
tensorflow==2.5.0